chore: reorganizacion tests
