+++ .npmignore
@@ -4,3 +4,5 @@
 .github
 old
 coverage
+.idea
+.vscode
